requirements.txt
flask
os
werkzeug.utils 
re
cv2
numpy 
pdf2image 
pytesseract
fitz
json
pandas
gunicorn
waitress